Refactor signature help logic into dedicated classes
Moved signature help computation from JavaCodeEditorPane to new SignatureHelpProvider interface and JdtSignatureHelpProvider implementation. Added SignatureHelp and ParameterInfo records for immutable signature metadata. JavaCodeEditorPane now delegates signature help requests to the provider, improving separation of concerns and maintainability.
